Rename tensor4all-matrixci to matrixci
This is an internal algorithm crate, not part of the public API.
