+++ Cargo.toml
@@ -4,7 +4,7 @@ members = [
     "crates/tensor4all-itensorlike",
     "crates/tensor4all-treetn",
     "crates/quanticsgrids",
-    "crates/tensor4all-matrixci",
+    "crates/matrixci",
     "crates/tensor4all-simpletensortrain",
     "crates/tensor4all-tensorci",
     "crates/tensor4all-capi",
