+++ crates/matrixci/src/error.rs
