+++ crates/tensor4all-core/Cargo.toml
@@ -12,7 +12,7 @@ backend-faer = ["mdarray-linalg-faer"]
 backend-lapack = ["mdarray-linalg-lapack"]
 
 [dependencies]
-tensor4all-matrixci = { path = "../tensor4all-matrixci" }
+matrixci = { path = "../matrixci" }
 num-complex.workspace = true
 num-traits.workspace = true
 mdarray.workspace = true
