+++ crates/tensor4all-simpletensortrain/Cargo.toml
@@ -27,7 +27,7 @@ mdarray-linalg.workspace = true
 mdarray-linalg-faer = { workspace = true, optional = true }
 mdarray-linalg-lapack = { workspace = true, optional = true }
 faer-traits.workspace = true
-tensor4all-matrixci = { path = "../tensor4all-matrixci" }
+matrixci = { path = "../matrixci" }
 tensor4all-core = { path = "../tensor4all-core" }
 
 [dev-dependencies]
