+++ crates/tensor4all-simpletensortrain/src/mpo/factorize.rs
@@ -128,7 +128,7 @@ where
         FactorizeMethod::RSVD => factorize_rsvd(matrix, options),
         FactorizeMethod::LU | FactorizeMethod::CI => {
             // For LU/CI, fall back to SVD for now
-            // Full LU/CI support requires tensor4all_matrixci::Scalar trait
+            // Full LU/CI support requires matrixci::Scalar trait
             factorize_svd(matrix, options)
         }
     }
@@ -290,24 +290,24 @@ where
 
 /// Factorize using LU decomposition
 ///
-/// This function requires the tensor4all_matrixci::Scalar trait.
+/// This function requires the matrixci::Scalar trait.
 /// Use this directly when you need LU-based factorization.
 pub fn factorize_lu<T: SVDScalar>(
     matrix: &Matrix2<T>,
     options: &FactorizeOptions,
 ) -> Result<FactorizeResult<T>>
 where
-    T: tensor4all_matrixci::util::Scalar,
+    T: matrixci::util::Scalar,
     <T as ComplexFloat>::Real: Into<f64>,
 {
-    use tensor4all_matrixci::{AbstractMatrixCI, MatrixLUCI, RrLUOptions};
+    use matrixci::{AbstractMatrixCI, MatrixLUCI, RrLUOptions};
 
     let m = matrix.dim(0);
     let n = matrix.dim(1);
 
     // Convert DTensor to matrixci::Matrix (temporary until matrixci migration)
-    let mut mat_ci: tensor4all_matrixci::util::Matrix<T> =
-        tensor4all_matrixci::util::zeros(m, n);
+    let mut mat_ci: matrixci::util::Matrix<T> =
+        matrixci::util::zeros(m, n);
     for i in 0..m {
         for j in 0..n {
             mat_ci[[i, j]] = matrix[[i, j]];
@@ -327,8 +327,8 @@ where
     let rank = luci.rank().max(1);
 
     // Convert back to DTensor
-    let left_m = tensor4all_matrixci::util::nrows(&left_ci);
-    let left_n = tensor4all_matrixci::util::ncols(&left_ci);
+    let left_m = matrixci::util::nrows(&left_ci);
+    let left_n = matrixci::util::ncols(&left_ci);
     let mut left: Matrix2<T> = matrix2_zeros(left_m, left_n);
     for i in 0..left_m {
         for j in 0..left_n {
@@ -336,8 +336,8 @@ where
         }
     }
 
-    let right_m = tensor4all_matrixci::util::nrows(&right_ci);
-    let right_n = tensor4all_matrixci::util::ncols(&right_ci);
+    let right_m = matrixci::util::nrows(&right_ci);
+    let right_n = matrixci::util::ncols(&right_ci);
     let mut right: Matrix2<T> = matrix2_zeros(right_m, right_n);
     for i in 0..right_m {
         for j in 0..right_n {
@@ -355,14 +355,14 @@ where
 
 /// Factorize using Cross Interpolation
 ///
-/// This function requires the tensor4all_matrixci::Scalar trait.
+/// This function requires the matrixci::Scalar trait.
 /// Use this directly when you need CI-based factorization.
 pub fn factorize_ci<T: SVDScalar>(
     matrix: &Matrix2<T>,
     options: &FactorizeOptions,
 ) -> Result<FactorizeResult<T>>
 where
-    T: tensor4all_matrixci::util::Scalar,
+    T: matrixci::util::Scalar,
     <T as ComplexFloat>::Real: Into<f64>,
 {
     // CI uses the same LUCI implementation as LU
