+++ crates/tensor4all-tensorci/Cargo.toml
@@ -13,7 +13,7 @@ num-traits.workspace = true
 anyhow.workspace = true
 thiserror.workspace = true
 rand.workspace = true
-tensor4all-matrixci = { path = "../tensor4all-matrixci" }
+matrixci = { path = "../matrixci" }
 tensor4all-simpletensortrain = { path = "../tensor4all-simpletensortrain" }
 
 [dev-dependencies]
